#136 feat: 휠 옵션 하나씩만 고를 수 있도록 구현
