+++ frontend/src/components/Option/OptionCardList/index.tsx
@@ -1,17 +1,18 @@
 import { useState, useEffect, MouseEvent } from 'react';
 
-import { useOutletContext } from 'react-router-dom';
+import { useLocation } from 'react-router-dom';
 
-import { OptionCardDataProps, OptionContextProviderProps } from '@/types/option';
+import { OptionCardDataProps } from '@/types/option';
 import { isIndexLargeThanZero, isIndexSmallThanMaxIndex } from '@/utils';
 import { OPTION_CARD_LIST_LENGTH } from '@/constants';
 
+import { OptionCard } from './OptionCard';
 import { PrevButton } from '@/components/common/PrevButton';
 import { NextButton } from '@/components/common/NextButton';
 
 import * as Styled from './style';
 
-interface OptionCardListProps {
+interface Props {
   selectedIndex: number;
   cardListIndex: number;
   data: OptionCardDataProps[];
@@ -19,28 +20,21 @@ interface OptionCardListProps {
   onClickArrowButton: (index: number, length: number) => void;
 }
 
-interface OptionCardProps {
-  index: number;
-  name: string;
-  additionalPrice: number;
-  imageUrl: string;
-  subOptionNames: string[];
-  isCardActive: boolean;
-  onClickCard: (index: number, event: MouseEvent<HTMLDivElement>) => void;
-}
-
-interface OptionCardHoverProps {
-  subOptionNames: string[];
-}
-
-export function OptionCardList({
-  selectedIndex,
-  cardListIndex,
-  data,
-  onClickCard,
-  onClickArrowButton,
-}: OptionCardListProps) {
+export function OptionCardList({ selectedIndex, cardListIndex, data, onClickCard, onClickArrowButton }: Props) {
   const [cardList, setCardList] = useState<OptionCardDataProps[]>([]);
+  const [selectedCount, setSelectedCount] = useState(0);
+  const [lastAddedIndex, setLastAddedIndex] = useState<number | null>(null);
+
+  const { pathname } = useLocation();
+
+  function handleSelectedCount(step: number, index: number) {
+    setSelectedCount(prev => prev + step);
+    setLastAddedIndex(index);
+  }
+
+  function isWheelSelected() {
+    return pathname === '/option/n-performance' && selectedCount === 1;
+  }
 
   useEffect(() => {
     const startIndex = cardListIndex * OPTION_CARD_LIST_LENGTH;
@@ -61,6 +55,7 @@ export function OptionCardList({
       />
       {cardList.map(({ index, name, additionalPrice, imageUrl, subOptionNames }) => (
         <OptionCard
+          isBlur={isWheelSelected() && index !== lastAddedIndex}
           index={index}
           name={name}
           additionalPrice={additionalPrice}
@@ -68,6 +63,7 @@ export function OptionCardList({
           subOptionNames={subOptionNames}
           isCardActive={index === selectedIndex}
           onClickCard={onClickCard}
+          onChangeCount={handleSelectedCount}
           key={index}
         />
       ))}
@@ -78,69 +74,5 @@ export function OptionCardList({
         isShow={isIndexSmallThanMaxIndex(cardListIndex, data.length)}
       />
     </Styled.Container>
-  );
-}
-
-function OptionCard({
-  index,
-  name,
-  additionalPrice,
-  imageUrl,
-  subOptionNames,
-  isCardActive,
-  onClickCard,
-}: OptionCardProps) {
-  const [isButtonActive, setIsButtonActive] = useState(false);
-  const [isHover, setIsHover] = useState(false);
-
-  const { trim, addOption, removeOption } = useOutletContext<OptionContextProviderProps>();
-
-  function handleClickButton() {
-    setIsButtonActive(prev => !prev);
-    isButtonActive
-      ? removeOption(name)
-      : addOption({ name, price: additionalPrice, imageUrl, subOptions: subOptionNames });
-  }
-
-  function handleHoverCard(isHover: boolean) {
-    setIsHover(isHover);
-  }
-
-  useEffect(() => {
-    const isOptionIncluded = trim.options.some(option => option.name === name);
-    setIsButtonActive(isOptionIncluded);
-  }, [name]);
-
-  return (
-    <Styled.OptionCard isCardActive={isCardActive} onClick={event => onClickCard(index, event)}>
-      <Styled.Image src={imageUrl} />
-      <Styled.DescriptionWrapper>
-        <Styled.Text isCardActive={isCardActive}>{name}</Styled.Text>
-        <Styled.Text isCardActive={isCardActive}>+{additionalPrice.toLocaleString()}원</Styled.Text>
-        <Styled.ButtonBox>
-          <Styled.Button isButtonActive={isButtonActive} onClick={handleClickButton}>
-            {isButtonActive ? '추가 완료' : '추가하기'}
-          </Styled.Button>
-        </Styled.ButtonBox>
-        {isButtonActive && <Styled.Icon src={'/src/assets/icons/icon_done.svg'} />}
-      </Styled.DescriptionWrapper>
-      {isHover && <OptionCardHover subOptionNames={subOptionNames} />}
-      <Styled.OptionCardHoverArea
-        onMouseEnter={() => handleHoverCard(true)}
-        onMouseLeave={() => handleHoverCard(false)}
-      />
-    </Styled.OptionCard>
-  );
-}
-
-function OptionCardHover({ subOptionNames }: OptionCardHoverProps) {
-  return (
-    <Styled.OptionCardHover>
-      <Styled.DescriptionHoverWrapper>
-        {subOptionNames.map((name, index) => (
-          <Styled.DescriptionHover key={index}>· {name}</Styled.DescriptionHover>
-        ))}
-      </Styled.DescriptionHoverWrapper>
-    </Styled.OptionCardHover>
   );
 }
